Implementación NYTimesServiceImpl
